Refactor visualisation component
