+++ designer/client/components/visualisation/index.ts
@@ -1,0 +1,1 @@
+export { Visualisation } from "./Visualisation";
