+++ designer/client/components/visualisation/minimap.scss
@@ -1,0 +1,113 @@
+@import "../../index.scss";
+
+.visualisation {
+  position: relative;
+  width: 100vw;
+  height: 100vh;
+
+  &__pages-wrapper {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    overflow: scroll;
+  }
+
+  .page {
+    position: absolute;
+    width: 240px;
+    margin-bottom: 10px;
+    box-sizing: border-box;
+    border-top: solid 2px #000;
+    border-right: solid 2px #000;
+    border-bottom: solid 2px #000;
+    border-left: solid 2px #000;
+
+    &--selected {
+      -webkit-box-shadow: 0px 0px 0px 6px $color-orange;
+      -moz-box-shadow: 0px 0px 0px 6px $color-orange;
+      box-shadow: 0px 0px 0px 6px $color-orange;
+    }
+
+    &:hover {
+      -webkit-box-shadow: 0px 0px 0px 6px $color-orange;
+      -moz-box-shadow: 0px 0px 0px 6px $color-orange;
+      box-shadow: 0px 0px 0px 6px $color-orange;
+    }
+
+    &__heading {
+      position: relative;
+      min-height: 2.8em;
+      @extend .govuk-\!-padding-2;
+
+      h3 {
+        @extend .govuk-heading-s;
+      }
+
+      span {
+        @extend .govuk-caption-m, .govuk-\!-font-size-14;
+      }
+    }
+
+    &__actions {
+      display: flex;
+      flex-direction: column;
+
+      & > * {
+        width: 100%;
+        @include govuk-font(14);
+        @extend .govuk-\!-padding-top-1,
+          .govuk-\!-padding-bottom-1,
+          .govuk-\!-padding-left-2,
+          .govuk-\!-padding-right-2;
+        border: none;
+        box-shadow: none;
+        text-decoration: none;
+        text-align: left;
+        background-color: govuk-colour("black");
+        color: govuk-colour("white");
+
+        &:hover {
+          background-color: $color-orange;
+          color: govuk-colour("black");
+          cursor: pointer;
+        }
+      }
+
+      & > *:not(:last-child) {
+        border-bottom: 1px solid #ffffff;
+      }
+    }
+
+    h3.govuk-heading-s {
+      margin-bottom: 0;
+    }
+  }
+
+  .notification {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    background-color: white;
+  }
+
+  .minimap {
+    position: absolute;
+    right: 0;
+    bottom: 0;
+    z-index: 998;
+    background: lightgray;
+    padding: 5px;
+    max-width: 50%;
+    max-height: 25%;
+    overflow: auto;
+    border-top: solid 1px #000;
+    border-left: solid 1px #000;
+
+    polyline {
+      stroke-width: 1;
+    }
+    rect:hover {
+      fill: $color-orange;
+    }
+  }
+}
